refactor(tracing): Adjust the API of JUnitTestCaseSorter#3033
